Gui: refactor AnimImage
