+++ applications/services/gui/modules/anim_image.c
@@ -1,9 +1,6 @@
-#include "anim_image.h"
+#include "anim_image_i.h"
 
 #include <furi/furi.h>
-
-#include <gui/widget_i.h>
-#include <storage/storage.h>
 
 #include <assets/assets_images.h>
 
@@ -28,28 +25,6 @@ typedef struct {
 static_assert(
     sizeof(AnimImageFileHeader) == 7 * sizeof(uint32_t),
     "Incorrect size of AnimImageFileHeader");
-
-typedef struct {
-    uint32_t begin_idx;
-    uint32_t end_idx;
-    bool loop;
-} AnimImageRange;
-
-struct AnimImage {
-    Widget base;
-    lv_obj_t* canvas;
-    lv_timer_t* timer;
-    uint8_t* canvas_buf;
-    File* file;
-    uint32_t frame_count;
-    size_t frame_size;
-    uint32_t current_idx;
-    AnimImageRange current_range;
-    AnimImageRange waiting_range;
-    bool has_waiting_range;
-};
-
-const lv_obj_class_t anim_image_lvgl_class;
 
 // Function prototypes
